+++ src/main/java/com/ipiecoles/java/audio/repository/AlbumRepository.java
@@ -3,7 +3,7 @@ package com.ipiecoles.java.audio.repository;
 import com.ipiecoles.java.audio.model.Album;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface AlbumRepository extends PagingAndSortingRepository<Album, Long> {
+public interface AlbumRepository extends PagingAndSortingRepository<Album, Integer> {
 
     // Méthodes
